+++ CHANGELOG.md
@@ -12,6 +12,9 @@ The format is based on [Keep a Changelog][chg] and this project adheres to
   - Add `AWS.Lambda.Events.EventBridge.SSM.ParameterStoreChange` type
     for parsing AWS Systems Manager Parameter Store Change events
     delivered via AWS EventBridge.
+  - When the runtime encounters an error, write it out to the Cloudwatch logs
+    (via stderr), and do so in a format that can be guaranteed and later
+    extended.
 
 ## `0.4.9` - 2022-02-28